+++ CHANGES.md
@@ -55,6 +55,7 @@
 - Switch back to single relay channel support.
 - Namespaces associated with task.
 - Fix initialization in W3C serialization library.
+- Support Fedora 35.
 ```
 
 ### v0.5.3